+++ sc/social/like/plugins/__init__.py
@@ -3,7 +3,7 @@ from sc.social.like.plugins.interfaces import IPlugin  # noqa
 
 
 class Plugin(object):
-    ''' Social Like Plugin Base Class'''
+    """ Social Like Plugin Base Class"""
 
     id = ''
     name = ''
@@ -19,3 +19,6 @@ class Plugin(object):
 
     def plugin(self):
         return 'plugin'
+
+    def link(self):
+        return 'link'