+++ sc/social/like/plugins/interfaces.py
@@ -3,26 +3,29 @@ from zope.interface import Interface
 
 
 class IPlugin(Interface):
-    ''' A Social Like Plugin '''
+    """ A Social Like Plugin """
 
     def config_view():
-        ''' Plugin configuration '''
+        """ Plugin configuration """
 
     def view():
-        ''' Plugin's browser view '''
+        """ Plugin's browser view """
 
     def metadata():
-        ''' Return metadata method '''
+        """ Return metadata method """
 
     def plugin():
-        ''' Return html render method '''
+        """ Return html render method """
+
+    def link():
+        """ Return html render method when "do_not_track" is active"""
 
 
 class IPluginView(Interface):
-    ''' A Social Like Plugin Browser View '''
+    """ A Social Like Plugin Browser View """
 
     def metadata():
-        ''' Return metadata method '''
+        """ Return metadata method """
 
     def plugin():
-        ''' Return html render method '''
+        """ Return html render method """